requirements-consumer.txt
kafka-python
flink